+++ Sources/Shared/Data Manager/Model/Social/SocialActivity.m
@@ -156,10 +156,10 @@
     
    NSMutableArray * links = [[NSMutableArray alloc] init];
     while ([string rangeOfString:@"<a href"].location!= NSNotFound) {
-        int beginTagLocation =[string rangeOfString:@"<a href"].location;
-        int contentTagLocation = [[string substringFromIndex:beginTagLocation] rangeOfString:@">"].location+1+beginTagLocation;
+        NSInteger beginTagLocation =[string rangeOfString:@"<a href"].location;
+        NSInteger contentTagLocation = [[string substringFromIndex:beginTagLocation] rangeOfString:@">"].location+1+beginTagLocation;
         
-        int endTagLocation =[[string substringFromIndex:beginTagLocation] rangeOfString:@"</a>"].location+1+beginTagLocation;
+        NSInteger endTagLocation =[[string substringFromIndex:beginTagLocation] rangeOfString:@"</a>"].location+1+beginTagLocation;
         
         NSString * link = [string substringWithRange:NSMakeRange(contentTagLocation, endTagLocation-contentTagLocation-1)];
         [links addObject:link];
